Use taglib instead of gstreamer to determine meta data for audio/video files
Taglib is a lot faster than gstreamer and redruces the time used for gather the meta data
a lot which is a good point to improve when dealing with large sets of media data.

Right now we're only feeding a small subset of meta data into our database. There will
be some rework soon to bring the meta data set in sync with the set used in legacy webOS.

Signed-off-by: Simon Busch <morphis@gravedo.de>